# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MDM Platform API - Mobile Device Management para Android TV
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Listar usuarios del tenant (solo admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Crear usuario en el tenant (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Eliminar usuario (solo admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
  /scheduled-tasks:
    get:
      operationId: getScheduledTasks
      tags: [scheduled-tasks]
      summary: Listar tareas programadas del tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledTask"
    post:
      operationId: createScheduledTask
      tags: [scheduled-tasks]
      summary: Crear tarea programada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduledTaskRequest"
      responses:
        "201":
          description: Tarea creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledTask"
  /scheduled-tasks/{id}:
    put:
      operationId: updateScheduledTask
      tags: [scheduled-tasks]
      summary: Actualizar tarea programada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduledTaskRequest"
      responses:
        "200":
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledTask"
    delete:
      operationId: deleteScheduledTask
      tags: [scheduled-tasks]
      summary: Eliminar tarea programada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator]
        tenantId:
          type: string
        tenantName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - tenantId
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, operator]
      required:
        - email
        - password
...
    ScheduledTask:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        deviceId:
          type: string
          nullable: true
        name:
          type: string
        cronExpression:
          type: string
        action:
          type: string
        actionParam:
          type: string
          nullable: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - cronExpression
        - action
        - enabled
        - createdAt
    CreateScheduledTaskRequest:
      type: object
      properties:
        name:
          type: string
        deviceId:
          type: string
        cronExpression:
          type: string
        action:
          type: string
        actionParam:
          type: string
        enabled:
          type: boolean
      required:
        - name
        - cronExpression
        - action